adding confirmExitHook back in, to be fixed in later commit
