+++ src/app/views/project/config/unsaved-config-confirmation.html
@@ -6,7 +6,7 @@
         Are you sure you want to leave this page? <strong>Any unsaved changes to the project config will be
         lost.</strong>
     </div>
-    <div layout="row" class="loading-overlay"  ng-show="$ctrl.loading" layout-align="center center">
+    <div layout="row" class="loading-overlay"  ng-show="vm.loading" layout-align="center center">
 	<md-progress-circular md-mode="indeterminate"></md-progress-circular>
     </div>
